refactor: apply rector autofixes (closures→arrow fns, modernization)
23 files modernized via composer rector-fix; resolves rector-check
failures by applying the dry-run-suggested transformations.

+++ interface/forms/upload_intake_form/view.php
@@ -181,7 +181,7 @@ if ($procedureReportId > 0) {
 
         $citationId = null;
         if ($key !== '' && !empty($bboxByField[$key])) {
-            $citationId = (int) array_shift($bboxByField[$key]);
+            $citationId = array_shift($bboxByField[$key]);
             if (empty($bboxByField[$key])) {
                 unset($bboxByField[$key]);
             }
